Startup the torrents in a separate thread
