+++ src/base/bittorrent/session.cpp
@@ -86,6 +86,7 @@
 #include "magneturi.h"
 #include "private/bandwidthscheduler.h"
 #include "private/filterparserthread.h"
+#include "private/resumedataloadingmanager.h"
 #include "private/resumedatasavingmanager.h"
 #include "private/statistics.h"
 #include "torrenthandle.h"
@@ -110,9 +111,6 @@ using namespace BitTorrent;
 
 namespace
 {
-    bool readFile(const QString &path, QByteArray &buf);
-    bool loadTorrentResumeData(const QByteArray &data, CreateTorrentParams &torrentParams, int &prio, MagnetUri &magnetUri);
-
     void torrentQueuePositionUp(const libt::torrent_handle &handle);
     void torrentQueuePositionDown(const libt::torrent_handle &handle);
     void torrentQueuePositionTop(const libt::torrent_handle &handle);
@@ -137,43 +135,6 @@ namespace
             result[i.key()] = i.value();
         return result;
     }
-
-    template <typename Entry>
-    QSet<QString> entryListToSetImpl(const Entry &entry)
-    {
-        Q_ASSERT(entry.type() == Entry::list_t);
-        QSet<QString> output;
-        for (int i = 0; i < entry.list_size(); ++i) {
-            const QString tag = QString::fromStdString(entry.list_string_value_at(i));
-            if (Session::isValidTag(tag))
-                output.insert(tag);
-            else
-                qWarning() << QString("Dropping invalid stored tag: %1").arg(tag);
-        }
-        return output;
-    }
-
-#if LIBTORRENT_VERSION_NUM < 10100
-    bool isList(const libt::lazy_entry *entry)
-    {
-        return entry && (entry->type() == libt::lazy_entry::list_t);
-    }
-
-    QSet<QString> entryListToSet(const libt::lazy_entry *entry)
-    {
-        return entry ? entryListToSetImpl(*entry) : QSet<QString>();
-    }
-#else
-    bool isList(const libt::bdecode_node &entry)
-    {
-        return entry.type() == libt::bdecode_node::list_t;
-    }
-
-    QSet<QString> entryListToSet(const libt::bdecode_node &entry)
-    {
-        return entryListToSetImpl(entry);
-    }
-#endif
 
     QString normalizePath(const QString &path)
     {
@@ -372,6 +333,7 @@ Session::Session(QObject *parent)
     , m_useProxy(false)
     , m_recentErroredTorrentsTimer(new QTimer(this))
     , m_startedCount(0)
+    , m_loadedCount(0)
     , m_started(false)
 {
     Logger *const logger = Logger::instance();
@@ -519,6 +481,7 @@ Session::Session(QObject *parent)
     m_resumeDataSavingManager = new ResumeDataSavingManager(m_resumeFolderPath);
     m_resumeDataSavingManager->moveToThread(m_ioThread);
     connect(m_ioThread, &QThread::finished, m_resumeDataSavingManager, &QObject::deleteLater);
+
     m_ioThread->start();
 
     // Regular saving of fastresume data
@@ -3810,85 +3773,44 @@ void Session::startUpTorrents()
 {
     qDebug("Resuming torrents...");
 
-    const QDir resumeDataDir(m_resumeFolderPath);
-    QStringList fastresumes = resumeDataDir.entryList(
-                QStringList(QLatin1String("*.fastresume")), QDir::Files, QDir::Unsorted);
-    Logger *const logger = Logger::instance();
+    ResumeDataLoadingManager *resumeDataLoadingManager = new ResumeDataLoadingManager(m_resumeFolderPath);
+    resumeDataLoadingManager->moveToThread(m_ioThread);
 
-    typedef struct
+
+
+    const auto startupTorrent = [this, resumeDataLoadingManager]()
     {
-        QString hash;
-        MagnetUri magnetUri;
-        CreateTorrentParams addTorrentData;
-        QByteArray data;
-    } TorrentResumeData;
+        int resumedTorrentsCount = 0;
+        QVector<ResumeDataLoadingManager::TorrentResumeData> resumeData;
+        resumeDataLoadingManager->getResumeData(resumeData);
 
-    int resumedTorrentsCount = 0;
-    const auto startupTorrent = [this, logger, &resumeDataDir, &resumedTorrentsCount](const TorrentResumeData &params)
-    {
-        QString filePath = resumeDataDir.filePath(QString("%1.torrent").arg(params.hash));
-        qDebug() << "Starting up torrent" << params.hash << "...";
-        ++m_startedCount;
-        if (!addTorrent_impl(params.addTorrentData, params.magnetUri, TorrentInfo::loadFromFile(filePath), params.data))
-            logger->addMessage(tr("Unable to resume torrent '%1'.", "e.g: Unable to resume torrent 'hash'.")
-                               .arg(params.hash), Log::CRITICAL);
+        for (const auto rd : resumeData) {
+            qDebug() << "Starting up torrent" << rd.hash << "...";
+            if (!addTorrent_impl(rd.torrentParams, rd.magnetUri, rd.info, rd.data))
+                LogMsg(tr("Unable to resume torrent '%1'.", "e.g: Unable to resume torrent 'hash'.")
+                                   .arg(rd.hash), Log::CRITICAL);
 
-        // process add torrent messages before message queue overflow
-        if ((resumedTorrentsCount % 100) == 0) readAlerts();
+            // process add torrent messages before message queue overflow
+            if ((resumedTorrentsCount % 100) == 0) readAlerts();
 
-        ++resumedTorrentsCount;
+            ++resumedTorrentsCount;
+        }
     };
 
-    qDebug("Starting up torrents");
-    qDebug("Queue size: %d", fastresumes.size());
-    // Resume downloads
-    QMap<int, TorrentResumeData> queuedResumeData;
-    int nextQueuePosition = 1;
-    int numOfRemappedFiles = 0;
-    const QRegularExpression rx(QLatin1String("^([A-Fa-f0-9]{40})\\.fastresume$"));
-    foreach (const QString &fastresumeName, fastresumes) {
-        const QRegularExpressionMatch rxMatch = rx.match(fastresumeName);
-        if (!rxMatch.hasMatch()) continue;
+    const auto torrentsLoaded = [this](int loadedCount)
+    {
+        m_loadedCount = loadedCount;
+        if (m_startedCount == m_loadedCount)
+            handleStartUpFinished();
+    };
 
-        QString hash = rxMatch.captured(1);
-        QString fastresumePath = resumeDataDir.absoluteFilePath(fastresumeName);
-        QByteArray data;
-        CreateTorrentParams torrentParams;
-        MagnetUri magnetUri;
-        int queuePosition;
-        if (readFile(fastresumePath, data) && loadTorrentResumeData(data, torrentParams, queuePosition, magnetUri)) {
-            if (queuePosition <= nextQueuePosition) {
-                startupTorrent({ hash, magnetUri, torrentParams, data });
+    connect(resumeDataLoadingManager, &ResumeDataLoadingManager::resumeDataReady, this, startupTorrent);
+    connect(resumeDataLoadingManager, &ResumeDataLoadingManager::loadingFinished, this, torrentsLoaded);
+    connect(resumeDataLoadingManager, &ResumeDataLoadingManager::loadingFinished, resumeDataLoadingManager, &QObject::deleteLater);
+    connect(m_ioThread, &QThread::finished, resumeDataLoadingManager, &QObject::deleteLater);
 
-                if (queuePosition == nextQueuePosition) {
-                    ++nextQueuePosition;
-                    while (queuedResumeData.contains(nextQueuePosition)) {
-                        startupTorrent(queuedResumeData.take(nextQueuePosition));
-                        ++nextQueuePosition;
-                    }
-                }
-            }
-            else {
-                int q = queuePosition;
-                for (; queuedResumeData.contains(q); ++q) {
-                }
-                if (q != queuePosition) {
-                    ++numOfRemappedFiles;
-                }
-                queuedResumeData[q] = {hash, magnetUri, torrentParams, data};
-            }
-        }
-    }
 
-    if (numOfRemappedFiles > 0) {
-        logger->addMessage(
-            QString(tr("Queue positions were corrected in %1 resume files")).arg(numOfRemappedFiles),
-            Log::CRITICAL);
-    }
-
-    // starting up downloading torrents (queue position > 0)
-    foreach (const TorrentResumeData &torrentResumeData, queuedResumeData)
-        startupTorrent(torrentResumeData);
+    QMetaObject::invokeMethod(resumeDataLoadingManager, "loadTorrents");
 }
 
 quint64 Session::getAlltimeDL() const
@@ -4134,7 +4056,7 @@ void Session::handleAddTorrentAlert(libt::add_torrent_alert *p)
 {
     if (p->error) {
         qDebug("/!\\ Error: Failed to add torrent!");
-        if (!m_started) --m_startedCount;
+        if (!m_started) ++m_startedCount;
         QString msg = QString::fromStdString(p->message());
         Logger::instance()->addMessage(tr("Couldn't add torrent. Reason: %1").arg(msg), Log::WARNING);
         emit addTorrentFailed(msg);
@@ -4147,8 +4069,8 @@ void Session::handleAddTorrentAlert(libt::add_torrent_alert *p)
 void Session::handleTorrentAddedAlert()
 {
     if (m_started) return;
-    --m_startedCount;
-    if (m_startedCount == 0)
+    ++m_startedCount;
+    if (m_startedCount == m_loadedCount)
         handleStartUpFinished();
 }
 
@@ -4506,71 +4428,6 @@ void Session::handleStateUpdateAlert(libt::state_update_alert *p)
 
 namespace
 {
-    bool readFile(const QString &path, QByteArray &buf)
-    {
-        QFile file(path);
-        if (!file.open(QIODevice::ReadOnly)) {
-            qDebug("Cannot read file %s: %s", qUtf8Printable(path), qUtf8Printable(file.errorString()));
-            return false;
-        }
-
-        buf = file.readAll();
-        return true;
-    }
-
-    bool loadTorrentResumeData(const QByteArray &data, CreateTorrentParams &torrentParams, int &prio,  MagnetUri &magnetUri)
-    {
-        torrentParams = CreateTorrentParams();
-        torrentParams.restored = true;
-        torrentParams.skipChecking = false;
-
-        libt::error_code ec;
-#if LIBTORRENT_VERSION_NUM < 10100
-        libt::lazy_entry fast;
-        libt::lazy_bdecode(data.constData(), data.constData() + data.size(), fast, ec);
-        if (ec || (fast.type() != libt::lazy_entry::dict_t)) return false;
-#else
-        libt::bdecode_node fast;
-        libt::bdecode(data.constData(), data.constData() + data.size(), fast, ec);
-        if (ec || (fast.type() != libt::bdecode_node::dict_t)) return false;
-#endif
-
-        torrentParams.savePath = Profile::instance().fromPortablePath(
-            Utils::Fs::fromNativePath(QString::fromStdString(fast.dict_find_string_value("qBt-savePath"))));
-
-        std::string ratioLimitString = fast.dict_find_string_value("qBt-ratioLimit");
-        if (ratioLimitString.empty())
-            torrentParams.ratioLimit = fast.dict_find_int_value("qBt-ratioLimit", TorrentHandle::USE_GLOBAL_RATIO * 1000) / 1000.0;
-        else
-            torrentParams.ratioLimit = QString::fromStdString(ratioLimitString).toDouble();
-        torrentParams.seedingTimeLimit = fast.dict_find_int_value("qBt-seedingTimeLimit", TorrentHandle::USE_GLOBAL_SEEDING_TIME);
-        // **************************************************************************************
-        // Workaround to convert legacy label to category
-        // TODO: Should be removed in future
-        torrentParams.category = QString::fromStdString(fast.dict_find_string_value("qBt-label"));
-        if (torrentParams.category.isEmpty())
-        // **************************************************************************************
-            torrentParams.category = QString::fromStdString(fast.dict_find_string_value("qBt-category"));
-        // auto because the return type depends on the #if above.
-        const auto tagsEntry = fast.dict_find_list("qBt-tags");
-        if (isList(tagsEntry))
-            torrentParams.tags = entryListToSet(tagsEntry);
-        torrentParams.name = QString::fromStdString(fast.dict_find_string_value("qBt-name"));
-        torrentParams.hasSeedStatus = fast.dict_find_int_value("qBt-seedStatus");
-        torrentParams.disableTempPath = fast.dict_find_int_value("qBt-tempPathDisabled");
-        torrentParams.hasRootFolder = fast.dict_find_int_value("qBt-hasRootFolder");
-
-        magnetUri = MagnetUri(QString::fromStdString(fast.dict_find_string_value("qBt-magnetUri")));
-        torrentParams.paused = fast.dict_find_int_value("qBt-paused");
-        torrentParams.forced = fast.dict_find_int_value("qBt-forced");
-        torrentParams.firstLastPiecePriority = fast.dict_find_int_value("qBt-firstLastPiecePriority");
-        torrentParams.sequential = fast.dict_find_int_value("qBt-sequential");
-
-        prio = fast.dict_find_int_value("qBt-queuePosition");
-
-        return true;
-    }
-
     void torrentQueuePositionUp(const libt::torrent_handle &handle)
     {
         try {
